Add support to 'listen' for tags
This adds support to break the `read_passive_target` function up into
two parts if desired. First tell the PN532 to listen for tags, and then
explicitly ask to read the UID from the tag. This can be useful when
trying to simultaneously scan tags and run timing sensitive code, like
updating neopixels, and don't want to wait for the timeout delay of
`read_passive_target`.